more main namespace deprecation
